% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_ts_plot.R
\name{agg_ts_plot}
\alias{agg_ts_plot}
\title{Draw 2 time-series plots of aggregated data for original and robust weights.}
\usage{
agg_ts_plot(
  robust_estimate,
  T_0,
  start_year = 1,
  folder = NULL,
  save_pdf = FALSE,
  height = 9,
  width = 16
)
}
\arguments{
\item{robust_estimate}{The object to use for building the plot.}

\item{T_0}{The number of periods used for constructing the weights.}

\item{start_year}{The starting year in the plot's x-axis.}

\item{folder}{The folder where the plots are saved. NULL stands for the working directory.}

\item{save_pdf}{If True, the 2 time-series plots are saved.}

\item{height}{The height of the plot.}

\item{width}{The width of the plot.}
}
\description{
Solid lines represent aggregated Y_t and W_t. Dashed ones represent OLS predictions of the aggregate data with the instrument.
Blue and red dashed lines in the robust weights plot stand for learning and estimation periods respectively.
}
