% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{basic_analysis}
\alias{basic_analysis}
\title{Calculate the necessary.}
\usage{
basic_analysis(Y_mat, W_mat, Z, X_unit, psi, D, T_0, seed = 1234)
}
\arguments{
\item{Y_mat}{an nxT matrix of outcome.}

\item{W_mat}{an nxT matrix of endogenous treatment.}

\item{Z}{an nx1 vector of an aggregate instrument.}

\item{X_unit}{unit covariate.}

\item{psi}{aggregate covariates.}

\item{D}{an nx1 vector of exposure.}

\item{T_0}{the size of the learning period.}

\item{seed}{a seed to set.}
}
\value{
A list with weights, coefficients, standard errors, delta and pi and aggregated W, Y (observed and predicted).
}
\description{
Calculate the necessary.
}
