% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_simulation.R
\name{basic_simulation}
\alias{basic_simulation}
\title{Calculate the necessary for one simulation.}
\usage{
basic_simulation(
  F_mat_W,
  L_mat_W,
  F_mat_Y,
  L_mat_Y,
  cov_mat_fs,
  cov_mat_rf,
  pi_unit,
  theta_w,
  theta_y,
  tau,
  no_noise,
  rho_agg,
  rho_cross,
  T_0,
  Z_fit,
  test = FALSE,
  S = 300
)
}
\arguments{
\item{F_mat_W}{a nxT two-way fixed effects matrix for W (endogenous treatment).}

\item{L_mat_W}{a nxT generalized fixed effects (low-rank) matrix for W (endogenous treatment).}

\item{F_mat_Y}{a nxT two-way fixed effects matrix for Y (outcome).}

\item{L_mat_Y}{a nxT generalized fixed effects (low-rank) matrix for Y (outcome).}

\item{cov_mat_fs}{a covarience matrix for W (endogenous treatment) noise.}

\item{cov_mat_rf}{a covarience matrix for a part of Y (outcome) noise.}

\item{pi_unit}{a nx1 vector of exposure.}

\item{theta_w}{an exposure to unobserved confounder for W (endogenous treatment).}

\item{theta_y}{an exposure to unobserved confounder for Y (outcome).}

\item{tau}{the original point estimate of tau. It is used for data generation.}

\item{no_noise}{If true, the noise is not added. We do not add noise for table 3.}

\item{rho_agg}{the share of Z_t during H_t generation.}

\item{rho_cross}{the share of W (endogenous treatment) noise in Y noise.}

\item{T_0}{the size of the learning period.}

\item{Z_fit}{the fitted ARIMA model to the data Z (instrument).}

\item{test}{If true, compute the tests for the coverage rates.}

\item{S}{if 'test' == TRUE, the number of simulations for standard error estimation.}
}
\value{
A list with simulation results.
}
\description{
Calculate the necessary for one simulation.
}
