% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biased var and cov.R
\name{cov_biased}
\alias{cov_biased}
\title{Biased covarience}
\usage{
cov_biased(x, y)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{y}{`NULL` (default) or a vector, matrix or data frame with compatible dimensions to x. The default is equivalent to y = x (but more efficient).}
}
\description{
The denominator is n.
}
