% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data preparation.R
\name{data_preparation}
\alias{data_preparation}
\title{Data preparation for simulations.}
\usage{
data_preparation(data_mat, rank)
}
\arguments{
\item{data_mat}{a matrix.}

\item{rank}{the rank of the low-rank matrix approximation with SVD.}
}
\value{
A list with generalized fixed effects (low-rank matrix), fixed effects matrix ,  ,unit fixed effects and .
}
\description{
Data preparation for simulations.
}
