% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format.robust_estimate}
\alias{format.robust_estimate}
\title{Format a robust estimator object.}
\usage{
\method{format}{robust_estimate.}(estimate)
}
\arguments{
\item{estimate}{The object to format.}
}
\description{
Format a robust estimator object.
}
