% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_function.R
\name{my_density_function}
\alias{my_density_function}
\title{Computations for density plots.}
\usage{
my_density_function(x, K, deg)
}
\arguments{
\item{x}{a vector of observed.}

\item{K}{number of domain splits.}

\item{deg}{degrees of freedom for natural cubic splines.}
}
\description{
Computations for density plots.
}
