% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_delta_scatterplot.R
\name{pi_delta_scatterplot}
\alias{pi_delta_scatterplot}
\title{Draw 2 scatterplots of reduced-form and first-stage coefficients for original and robust weights.}
\usage{
pi_delta_scatterplot(
  robust_estimate,
  folder = NULL,
  save_pdf = FALSE,
  height = 9,
  width = 9
)
}
\arguments{
\item{robust_estimate}{The object to use for building the plot.}

\item{folder}{The folder where the plots are saved. NULL stands for the working directory.}

\item{save_pdf}{If True, the 2 scatterplots are saved.}

\item{height}{The height of the plot.}

\item{width}{The width of the plot.}
}
\description{
Draw 2 scatterplots of reduced-form and first-stage coefficients for original and robust weights.
}
