% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_1.R
\name{plot_1}
\alias{plot_1}
\title{Draw 2 scatterplots of reduced-form and first-stage coefficients for original and robust weights.}
\usage{
plot_1(robust_estimate, folder = NULL, height = 9, width = 9)
}
\arguments{
\item{robust_estimate}{The object to use for building the plot.}

\item{folder}{The folder where the plots are saved. NULL stands for the working directory.}

\item{height}{The height of the plot.}

\item{width}{The width of the plot.}
}
\description{
Draw 2 scatterplots of reduced-form and first-stage coefficients for original and robust weights.
}
