% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2.R
\name{plot_2}
\alias{plot_2}
\title{Draw 2 time-series plots of aggregated data for original and robust weights.}
\usage{
plot_2(robust_estimate, T_0, start_year = 1, height = 9, width = 16)
}
\arguments{
\item{robust_estimate}{The object to use for building the plot.}

\item{T_0}{The number of periods used for constructing the weights.}

\item{start_year}{The starting year in the plot's x-axis.}

\item{height}{The height of the plot.}

\item{width}{The width of the plot.}

\item{folder}{The folder where the plots are saved. NULL stands for the working directory.}
}
\description{
Solid lines represent aggregated Y_t and W_t. Dashed ones represent OLS predictions of the aggregate data with the instrument.
Blue and red dashed lines in the robust weights plot stand for learning and estimation periods respectively.
}
