% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3.R
\name{plot_3}
\alias{plot_3}
\title{Draw a comparative scatterplot for original weights and robust weights.}
\usage{
plot_3(robust_estimate, folder = NULL, width = 9, height = 9)
}
\arguments{
\item{robust_estimate}{The object to use for building the plot.}

\item{folder}{The folder where the plots are saved. NULL stands for the working directory.}

\item{width}{The width of the plot.}

\item{height}{The height of the plot.}
}
\description{
Draw a comparative scatterplot for original weights and robust weights.
}
