% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.robust_estimate}
\alias{print.robust_estimate}
\title{Print a robust_estimate object}
\usage{
\method{print}{robust_estimate.}(estimate)
}
\arguments{
\item{x}{The object to print.}
}
\description{
Print a robust_estimate object
}
