% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show,robust_estimate-method}
\alias{show,robust_estimate-method}
\title{Show a robust_estimate object.}
\usage{
\S4method{show}{robust_estimate}(object)
}
\arguments{
\item{object}{The object to print.}
}
\description{
Show a robust_estimate object.
}
