% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_from_file.R
\name{sim_from_file}
\alias{sim_from_file}
\title{Present the simulation results from simulation file.}
\usage{
sim_from_file(
  file_folder = NULL,
  file_name = "simulation_result",
  plot_folder = NULL,
  K = 300,
  deg = 4,
  height = 9 * 0.75,
  width = 16 * 0.75
)
}
\arguments{
\item{file_folder}{The folder where the simulation file is located. NULL stands for the working directory.}

\item{file_name}{The name of the simulation file.}

\item{plot_folder}{The folder where the plot is saved. NULL stands for the working directory.}

\item{K}{number of domain splits. It is used in density plot.}

\item{deg}{degrees of freedom for natural cubic splines. It is used in density plot.}

\item{height}{the height of the plot.}

\item{width}{the width of the plot.}
}
\description{
Present the simulation results from simulation file.
}
