% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Perform simulations.}
\usage{
simulation(
  Y_mat_or,
  W_mat_or,
  Z,
  share_t = 1/3,
  share_rank = 1/3,
  rho_agg = 0.5,
  rho_theta_w = 0.2,
  rho_theta_y = 0.3,
  B = 1000,
  S = 300,
  test = FALSE,
  K = 300,
  deg = 4,
  height = 9 * 0.75,
  width = 16 * 0.75,
  folder = NULL,
  seed = 1234
)
}
\arguments{
\item{Y_mat_or}{an nxT matrix of outcome.}

\item{W_mat_or}{an nxT matrix of endogenous treatment.}

\item{Z}{n nx1 vector of an aggregate instrument.}

\item{share_t}{the share of the learning period.}

\item{share_rank}{the share of the rank of generalized fixed effects matrix in total rank.}

\item{rho_agg}{the share of Z_t during H_t generation.}

\item{rho_theta_w}{the exposure of W (endogenous treatment) to unobserved confounder.}

\item{rho_theta_y}{the exposure of Y (outcome) to unobserved confounder.}

\item{B}{number of replications for each design of simulation.}

\item{S}{if 'test' == TRUE, the number of simulations for standard error estimation.}

\item{test}{If true, compute the tests for the coverage rates.}

\item{K}{number of domain splits. It is used in density plot.}

\item{deg}{degrees of freedom for natural cubic splines. It is used in density plot.}

\item{height}{the height of the plot.}

\item{width}{the width of the plot.}

\item{folder}{The folder where the plots are saved. NULL stands for the working directory.}

\item{seed}{seed to set.}
}
\description{
Perform simulations.
}
