% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulations}
\alias{simulations}
\title{Perform the simulations and present the results.}
\usage{
simulations(
  Y_mat_or,
  W_mat_or,
  Z,
  share_t = 1/3,
  share_rank = 1/3,
  no_noise = FALSE,
  rho_agg = 0.5,
  rho_theta_w = 0.2,
  rho_theta_y = 0.3,
  B = 1000,
  S = 300,
  test = FALSE,
  return_table = FALSE,
  K = 300,
  deg = 4,
  height = 9 * 0.75,
  width = 16 * 0.75,
  save_pdf = FALSE,
  file_plot = "density_plot",
  draw_plot = TRUE,
  save_sim = FALSE,
  file_sim = "simulation_result",
  seed = 1234
)
}
\arguments{
\item{Y_mat_or}{an nxT matrix of outcome.}

\item{W_mat_or}{an nxT matrix of endogenous treatment.}

\item{Z}{n nx1 vector of an aggregate instrument.}

\item{share_t}{the share of the learning period.}

\item{share_rank}{the share of the rank of generalized fixed effects matrix in total rank.}

\item{no_noise}{If true, the noise is not added. We do not add noise for table 3.}

\item{rho_agg}{the share of Z_t during H_t generation.}

\item{rho_theta_w}{the exposure of W (endogenous treatment) to unobserved confounder.}

\item{rho_theta_y}{the exposure of Y (outcome) to unobserved confounder.}

\item{B}{number of replications for each design of simulation.}

\item{S}{if 'test' == TRUE, the number of simulations for standard error estimation.}

\item{test}{If true, compute the tests for the coverage rates.}

\item{return_table}{If true, instead of printing table(s) the function returns it(them).}

\item{K}{number of domain splits. It is used in density plot.}

\item{deg}{degrees of freedom for natural cubic splines. It is used in density plot.}

\item{height}{the height of the plot.}

\item{width}{the width of the plot.}

\item{save_pdf}{If true, the 2 density plots for Original and Robust estimators are saved.}

\item{file_plot}{If save_pdf is true, string giving the file path (including the file name).}

\item{draw_plot}{If draw_plot is true and save_pdf is false, the 2 density plots are drawn in one chart.}

\item{save_sim}{If true, the function saves the simulation.}

\item{file_sim}{If save_sim is true, string giving the file path (including the file name).}

\item{seed}{seed to set.}
}
\description{
Perform the simulations and present the results.
}
