% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_function.R
\name{weights_function}
\alias{weights_function}
\title{Calculate the weights}
\usage{
weights_function(
  Y_mat,
  W_mat,
  Z_agg,
  D_unit,
  unit_covariates = NULL,
  time_covariates = NULL,
  lambda = "basic"
)
}
\arguments{
\item{Y_mat}{an nxT matrix of outcome.}

\item{W_mat}{an nxT matrix of endogenous treatment.}

\item{Z_agg}{an nx1 vector of an aggregate instrument.}

\item{D_unit}{an nx1 vector of exposure.}

\item{unit_covariates}{unit covariate(s).}

\item{time_covariates}{time covariate(s).}

\item{lambda}{regularization hyperparameter.}
}
\value{
A list with weights.
}
\description{
Calculate the weights
}
