% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{basic_analysis}
\alias{basic_analysis}
\title{Calculate the necessary for the estimators.}
\usage{
basic_analysis(
  Y_mat,
  W_mat,
  Z,
  unit_covariates,
  time_covariates,
  add_const,
  D,
  T_0,
  seed = 1234
)
}
\arguments{
\item{Y_mat}{Y, the nxT matrix of outcome.}

\item{W_mat}{W, the nxT matrix of endogenous treatment.}

\item{Z}{Z, the nx1 vector of an aggregate instrument.}

\item{unit_covariates}{unit covariate(s).}

\item{time_covariates}{time covariate(s).}

\item{add_const}{If true, add constant for weights estimation.}

\item{D}{the nx1 vector of exposure.}

\item{T_0}{the size of the learning period.}

\item{seed}{a seed to set.}
}
\value{
A list with weights, coefficients, standard errors, delta and pi and aggregated W, Y (observed and predicted).
}
\description{
Calculate the necessary for the estimators.
}
