% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_delta_scatterplot.R
\name{pi_delta_scatterplot}
\alias{pi_delta_scatterplot}
\title{Draw 2 scatterplots of reduced-form and first-stage coefficients for original and robust weights.}
\usage{
pi_delta_scatterplot(
  robust_estimate,
  save_pdf = FALSE,
  file = "nak_points",
  height = 9,
  width = 9
)
}
\arguments{
\item{robust_estimate}{The object to use for building the plot.}

\item{save_pdf}{If True, the 2 scatterplots are saved and not displayed.}

\item{file}{If save_pdf is true, string giving the file path (including the file name). `_robust` and `_original` are added to the end of the file name for Robust and Original time-series respectively.}

\item{height}{The height of the plot.}

\item{width}{The width of the plot.}
}
\description{
Draw 2 scatterplots of reduced-form and first-stage coefficients for original and robust weights.
}
