% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_estimate.R
\name{robust_estimate}
\alias{robust_estimate}
\title{Compute the robust estimator for tau.}
\usage{
robust_estimate(
  Y_mat_or,
  W_mat_or,
  Z,
  unit_covariates = NULL,
  time_covariates = NULL,
  index_sub = NULL,
  T_0 = NULL,
  unit_names = NULL,
  time_column = TRUE,
  seed = 1234
)
}
\arguments{
\item{Y_mat_or}{an nxT matrix of outcome.}

\item{W_mat_or}{an nxT matrix of endogenous treatment.}

\item{Z}{an nx1 vector of an aggregate instrument.}

\item{index_sub}{an nx1 vector with binary coordinates. If one of the coordinates is TRUE, we include this state in the estimation.NULL stands for selecting all the units.}

\item{T_0}{the size of the learning period.}

\item{unit_names}{an nx1 vector with unit names. It is used only in plot_1 function.}

\item{time_column}{If true, the units are in the rows and the time periods are in the columns in Y_mat_or and W_mat_or.}

\item{seed, }{seed to set.}

\item{D}{an nx1 vector of exposure.}
}
\value{
A robust estimator with result, attached as an attribute. It is an output of basic_analysis function.
}
\description{
Compute the robust estimator for tau.
}
