% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_from_file.R
\name{sim_from_file}
\alias{sim_from_file}
\title{Present the simulation results from simulation file.}
\usage{
sim_from_file(
  sim_file = "simulation_result",
  plot_file = "density_plot",
  K = 300,
  deg = 4,
  height = 9 * 0.75,
  width = 16 * 0.75
)
}
\arguments{
\item{sim_file}{String giving the simulation file path (including the file name).}

\item{plot_file}{String giving the plot file path (including the file name).}

\item{K}{number of domain splits. It is used in density plot.}

\item{deg}{degrees of freedom for natural cubic splines. It is used in density plot.}

\item{height}{the height of the plot.}

\item{width}{the width of the plot.}
}
\description{
Present the simulation results from simulation file.
}
