% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_function.R
\name{weights_function}
\alias{weights_function}
\title{Calculate the original and robust weights}
\usage{
weights_function(
  Y_mat,
  W_mat,
  Z_agg,
  D_unit,
  X_unit = NULL,
  psi = NULL,
  lambda = "basic"
)
}
\arguments{
\item{Y_mat}{an nxT matrix of outcome.}

\item{W_mat}{an nxT matrix of endogenous treatment.}

\item{Z_agg}{an nx1 vector of an aggregate instrument.}

\item{D_unit}{an nx1 vector of exposure.}

\item{X_unit}{unit covariate.}

\item{psi}{aggregate covariates.}

\item{lambda}{regularization hyperparameter.}
}
\value{
A list with original and robust weight vectors.
}
\description{
Calculate the original and robust weights
}
